[Fix] YearSlider
