+++ src/components/main.tsx
@@ -5,6 +5,7 @@ import { LoadingBox } from "./loadingBox";
 import { MapComponent } from "./mapLibre";
 import { Panel } from "./panel";
 import { MainCanvas } from "./useThree/mainCanvas";
+import { YearSlider } from "./useThree/yearSlider";
 import type { MoonquakeData } from "@/type";
 import { Option, OptionConstants } from "@/type/option";
 import { fetchArtificialImpactCSV, fetchDeepMoonquakeCSV, fetchShallowMoonquakeCSV } from "@/utils/fetchMoonquakeCSV";
@@ -39,6 +40,7 @@ export const Main = () => {
     <Box w="100%" h="100svh" position="relative" overflow="none">
       <Header option={option} setOption={setOption} />
       <Panel choiceMoonquake={choiceMoonquake} />
+      {isMap || <YearSlider option={option} setOption={setOption} />}
 
       <LoadingBox loadingPageStep={loadingPageStep} />
