+++ src/components/useHeader/index.ts
@@ -1,3 +1,2 @@
 export { SwitchSetting } from "./switchSetting";
 export { TypeFilterSetting } from "./typeFilterSetting";
-export { YearSlider } from "./yearSlider";
